feat(params): Support "negative" sources and layers
This adds support for negative sources and layers as described in
https://github.com/pelias/pelias/issues/844.

It works by detecting if any sources or layers in the query parameters
are prefixed by '-'. In that case, the list of sources/layers to be
filtered against will _not_ include that source or layer.

Negative and positive layers can be combined in a query, with the
negative layers being applied _after_, and therefore taking priority
over, positive layers.

This is especially useful with aliases: something like
`layers=coarse,-country` would select all coarse layers except the
country layer.