fix(address layer filter): support negative sources and layers
The address layer filter is a performance optimization that excludes the
`address` layer from search results when the input doesn't look like an
address.

Logic already existed to make sure this optimization doesn't do harm:
for example it will not remove the address layer from results if the
address layer was explicitly specified in the `layers` parameter.

It also won't remove addresses from search results if doing so would
lead to the list of layers to search against being empty. This can
happen if `sources=openaddresses` (or any custom source that includes
only addresses) is specified too.

All this logic needs to know about _negative_ sources and layers.
Mainly, this code now needs to handle the fact that specifying negative
layers means the address layer might still be present and can be
removed. The code now takes no action if _positive_ layers are
specified, but if only _negative_ sources are specified then it still
evaluates if the `address` layer can be removed.

This helps maintain performance if something like `layers=-dependency`
is set, for example.

The other half of the change is in _how_ the address layer is removed.
Previously, whenever there were no sources or layer parameters set, this
code could set the `clean.layers` property itself, knowing that nothing
had set it before.

Now, because only sources and _positive_ layers are directly relevant to
how this code runs, the `clean.layers` value has to be modified, since
it might have been set as a result of _negative_ layers.